Add mutability switch to bindings and attributes
Make bindings and attributes immutable (throw exception) when they are
used for the first marshalling/unmarshalling operation, in order to
guarantee threadsafe operation. The mutability switch is added to the
'top level' binding definition that implements IModelWare.

Also, license information was added to a few test cases that were
missing them.