Add nil-support to unmarshalling of all relevant IBindings.
These are: Choice, ComplexType, Element (including subclasses Root and
Reference), MapRepeater, Repeater, Recursor, Sequence, SimpleFileType
and SimpleType.

A check is made if the nil-element is valid, meaning: it has no content.

Ignore has no nil-support, because it is ignoring everything and this
supersedes the nil-option. However, this implies that an xml-element
bound to an Ignore, misses the nil-validation when the xml-value is
nil.